AZ 1018: fix busy_port and busy_dist_port monitoring
Change the way busy port messages are filtered.

Old way: log all events up to a count.
New way: log all events for a unique port up to the count.

Note that the new way still counts all events, regardless of uniqueness,
to log the number of messages suppressed by the filter.

In the case of busy_dist_port events, we attempt to find the Erlang
node name that is on the far side of the port's TCP socket.  This
can help diagnose point-to-point network congestion problems.