Work-around for net_kernel:get_status() race condition
The node monitor race in `net_kernel:get_status()` is present
both in all (?) R14B releases as well as R15B.  It's not going to
be fixed anytime quickly.

So, the work-around involves even more knowledge of the internals
of `net_kernel`, but we were already breaking that barrier: in for
a penny, in for a dollar.

Passes basho_expect's regression_az1018 test.